+++ src/app/(none-layout)/auth/send-email/page.tsx
@@ -1,12 +1,14 @@
-import React from 'react';
+import React, { Suspense } from 'react';
 import VerifyEmailNotification from '@/components/signup/VerifyEmailNotification';
 import CommonPageLayout from '@/components/layout/CommonPageLayout';
 
 const page = () => {
   return (
-    <CommonPageLayout>
-      <VerifyEmailNotification />
-    </CommonPageLayout>
+    <Suspense fallback={<p>...loading</p>}>
+      <CommonPageLayout>
+        <VerifyEmailNotification />
+      </CommonPageLayout>
+    </Suspense>
   );
 };
